\name{btcol}
\alias{btcol}
\title{Auxilliary function computing factors.}
\usage{
btcol(col)
}
\arguments{
  \item{col}{Variables t,a and j to be combined}
}
\value{Coefficients need to be computed for the entire permutation matrix of indices,
this is the columnwise evaluation based on \code{t},\code{a} and \code{j}.}
\description{
Auxilliary function, that evaluates coefficents for elements of the indices matrix.
}
